feat: ダークモード切替ボタンをヘッダーに統合
ヘッダーの操作部に切り替えボタンを統合しました。
