+++ front/react-app/src/Molecules/HeaderLogins.tsx
@@ -11,8 +11,10 @@ import { useNavigate } from "react-router-dom";
 import { BsAlarm } from 'react-icons/bs';
 import { VscBell } from 'react-icons/vsc';
 import { MdLogout } from 'react-icons/md';
+import { DarkModeProps } from "../types/types";
+import { BsSun, BsMoon } from 'react-icons/bs';
 
-export const HerderLogins = (): JSX.Element => {
+export const HerderLogins = ({ isDark, toggleDarkClassForHtml } : DarkModeProps): JSX.Element => {
   const [showLoginModal, setShowLoginModal] = useState<boolean>(false);
   const [showSignUpModal, setShowSignUpModal] = useState<boolean>(false);
   // const [showAlarmModal, setShowAlarmModal] = useState<boolean>(false);
@@ -73,7 +75,7 @@ export const HerderLogins = (): JSX.Element => {
       { showSignUpModal ? <SignUpModal onClick={ hiddenSignUpModal } /> : <></> }
       { showAlarmFlag ? <AlarmModal onClick={ hiddenAlarmModal } /> : <></> }
       { loginFlag ? <>
-          <div className="border-2 border-gray-700 bg-sky-100 dark:bg-gray-600 p-1 h-full w-72 flex items-center justify-around rounded-lg">
+          <div className="border-2 border-gray-700 bg-sky-100 dark:bg-gray-600 p-1 h-full w-96 flex items-center justify-around rounded-lg">
             <button className="h-full w-32 text-lg font-semibold rounded-lg border-2 border-amber-400 dark:border-gray-400 bg-amber-100 dark:bg-gray-500 hover:bg-amber-200" onClick={ openAlarmModal } >
               アラーム<BsAlarm className="inline-block mb-1 ml-2" />
             </button>
@@ -85,6 +87,13 @@ export const HerderLogins = (): JSX.Element => {
                 通知<VscBell className="inline-block mb-1 ml-1" />
               </button>
             </div>
+            <button onClick={ toggleDarkClassForHtml } className="border-2 border-gray-400 bg-amber-100 dark:bg-inherit dark:border-gray-300 w-16 h-full rounded-md">
+              { isDark ? 
+                <BsSun className='text-black inline-block hover:cursor-pointer w-full h-6' />
+              :
+                <BsMoon className='inline-block hover:cursor-pointer w-full h-6' />
+              }
+            </button>
           </div>
         </> : <>
           <div className="border-2 border-gray-400 bg-sky-100 dark:bg-gray-600 rounded-md p-2 h-full w-96 flex items-center justify-around">
